requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
flask-login==0.6.3
flask-cors==4.0.0
flask-mail==0.9.1
python-dotenv==1.0.0
itsdangerous==2.1.2
pymssql==2.3.8
azure-storage-blob==12.19.0
openai==1.12.0
requests==2.31.0
pandas==2.1.4
beautifulsoup4==4.12.2
python-dateutil==2.8.2
pyjwt==2.8.0
gunicorn==21.2.0
pyotp==2.9.0
groq==0.4.1
tavily-python==0.3.0
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
waitress==2.1.2
